tests/wpt/tests/cookie-deprecation-label/cookie-deprecation-label-insecure-context.http.html

<!doctype html>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
  test(() => {
    assert_equals(navigator.cookieDeprecationLabel, undefined);
  }, 'Test navigator.cookieDeprecationLabel in insecure context');
</script>
